Vehicle Instants functionality written

+++ src/iimas/tum/models/Vehicle.java
@@ -5,23 +5,23 @@ import org.json.JSONObject;
 
 public class Vehicle {
 
-	protected int uid;
+	protected long uid;
 	protected int id;
 	protected int lineId;
 
-	Vehicle(Integer uniqueIdentifier, int identifier, int lineId) {
+	Vehicle(long uniqueIdentifier, int identifier, int lineId) {
 		this.uid = uniqueIdentifier;
 		this.id = identifier;
 		this.lineId = lineId;
 	}
 	
 	public Vehicle(JSONObject object) throws JSONException {
-		this(object.getInt("identifier"),
+		this(object.getLong("identifier"),
 				object.getInt("id"), 
 				object.getInt("lineId"));
 	}
 	
-	public int getUId() {
+	public long getUId() {
 		return this.uid;
 	}
